Created a very generic ApiCall factory to put placeholders into the GraphController. Also did some refactoring and created a '/private' folder
