+++ public/javascripts/Enums.js
@@ -1,4 +1,3 @@
-
 /*
 For enums, I would want the value to be an object instead of an enum. Comparisons would be at the very least slightly
 more robust - I wouldn't have any possibility of matching positively against a normal int or string value.
@@ -13,7 +12,7 @@ For graph generation, there is a need to maintain some sort of state to differen
 simplest example, a click is different when clicking on empty space (generating a node) versus clicking on a node itself
 (selecting a node).
  */
-var STATE = {
+const STATE = {
     IDLE : new enumValue('IDLE'),
     HOVEREMPTY : new enumValue('HOVEREMPTY'),
     HOVERNODE : new enumValue('HOVERNODE'),
@@ -23,8 +22,25 @@ var STATE = {
     CONTEXTMENU : new enumValue('CONTEXTMENU')
 };
 
-var TYPE = {
+const TYPE = {
     EMPTY : new enumValue('EMPTY'),
     NODE : new enumValue('NODE'),
     LINK : new enumValue('LINK')
+};
+
+const HTTP = {
+  GET : "GET",
+  POST : "POST",
+  PUT : "PUT",
+  DELETE : "DELETE"
+};
+
+const API_REQUEST = {
+  NONE_MADE: "ERR_NO_REQUEST_MADE",
+  NOT_DONE: "ERR_REQUEST_NOT_DONE"
+};
+
+const ENDPOINT = {
+  devo: "localhost:3000/api",
+  prod: ""
 };
