+++ public/javascripts/GraphState.js
@@ -90,142 +90,8 @@ $(document).ready(function(){
     setControls();
 });
 
-/*
-Create a new node at a given (X,Y) coordinate
- */
-function addNode(position) {
-    var uuid = uuidv4();
-    var nodeToAdd = {
-        id: uuid,
-        name: "Node " + uuid,
-        group: Math.floor(Math.random()*100)%5,
-        x: position.pageX,
-        y: position.pageY
-    };
-    data.nodes.push(nodeToAdd);
-    update();
-}
-
-/*
-Create a link between two nodes and add it to the graph
- */
-function addLink(fromNode, toNode) {
-    var linkToAdd = {
-        source: fromNode.id,
-        target: toNode.id,
-        value: 1
-    };
-    data.links.push(linkToAdd);
-    update();
-}
-
-/*
-Update each group of objects in the D3 force simulation.
- */
-function ticked() {
-    svg.selectAll("circle")
-        .attr("r", nodeRadius)
-        .style("fill", function(d) {return color(d.group)})
-        .style("stroke", "#424242")
-        .style("stroke-width", borderWidth)
-        .attr("cx", function (d) { return d.x; })
-        .attr("cy", function(d) { return d.y; });
-    svg.selectAll("line")
-        .style("stroke", "#aaa")
-        .style("fill", "#aaa")
-        .style("stroke-width", borderWidth)
-        .attr("x1", function(d) { return d.source.x; })
-        .attr("y1", function(d) { return d.source.y; })
-        .attr("x2", function(d) { return d.target.x; })
-        .attr("y2", function(d) { return d.target.y; });
-    svg.selectAll("text")
-        .attr("x", function(d) { return d.x; })
-        .attr("y", function (d) { return d.y - 1.2*nodeRadius; })
-        .style("font-family", "Lato")
-        .style("font-size", "12px")
-        .style("fill", "#333");
-}
-
-/*
-Drag methods to adjust the position of the node we are dragging, and also to update the state machine in case any other
-controls are activated while we do this.
- */
-function dragstarted(d) {
-    setState(STATE.DRAGNODE);
-    if (!d3.event.active) simulation.alphaTarget(0.3).restart()
-    d.fx = d.x;
-    d.fy = d.y;
-}
-function dragged(d) {
-    d.fx = d3.event.x;
-    d.fy = d3.event.y;
-}
-function dragended(d) {
-    setState(STATE.HOVERNODE);
-    d.fx = null;
-    d.fy = null;
-    if (!d3.event.active) simulation.alphaTarget(0);
-}
-
-/*
-Enable controls based on our state machine.
-
-TODO: Create a file for event handling in regards to controls. Refactor the shit out of this too.
-We should handle the events ourselves, at least w/ mousedown/up. The generalizations in jquery will hold us back.
- */
-function setControls() {
-    $("svg").on("click", function(event) {
-        switch(state) {
-            case STATE.IDLE:
-                setState(STATE.HOVEREMPTY);
-                setSelection({}, TYPE.EMPTY);
-                break;
-            case STATE.HOVEREMPTY:
-                addNode(event);
-                break;
-            case STATE.HOVERNODE:
-                // TODO: Select method
-                alert('Selecting Node!' + JSON.stringify(selection));
-                break;
-            case STATE.CONTEXTMENU:
-                // TODO: Select option or click out of context menu.
-                break;
-            default:
-                logDebug("No click handler for " +  state.value);
-                return;
-        }
-    });
-
-    $("svg").on("contextmenu", function(event) {
-        //Disable that default context menu shit
-        event.preventDefault();
-        switch (state) {
-            case STATE.HOVEREMPTY:
-                // TODO: Context Menu?
-                break;
-            case STATE.HOVERNODE:
-                if(getSpecialState() != STATE.LINKSTART) {
-                    setSpecialState(STATE.LINKSTART);
-                    setSpecialSelection(getSelection().object, getSelection().type);
-                    changeBGColor("#acc6d3");
-                } else {
-                    addLink(getSpecialSelection().object, getSelection().object);
-                    resetSpecialState();
-                    clearSpecialSelection();
-                    changeBGColor("#d3d3d3");
-                }
-                break;
-            default:
-                logDebug("No right mouse down handler for " + state.value);
-                return;
-        }
-    });
-}
-
-//Debug function I just need some visual feedback on linking ugh
-function changeBGColor(hex) {
-    $("#graphContainer").css("background-color", hex);
-}
+//////////////////////////////////////////
+//  State Management
 
 function getState(){ return state; }
 function getSpecialState(){ return specialState; }
