+++ routes/api.js
@@ -1,0 +1,9 @@
+var express = require('express');
+var router = express.Router();
+
+/* GET home page. */
+router.get('/', function(req, res, next) {
+  res.render('adventure', { title: 'Storied' });
+});
+
+module.exports = router;
