+++ views/adventure.pug
@@ -11,6 +11,7 @@ block footer
   script(src="/javascripts/Enums.js")
   script(src="/javascripts/Logger.js")
   script(src="/javascripts/GraphUpdater.js")
-  script(src="/javascripts/GraphManager.js")
+  script(src="/javascripts/GraphController.js")
+  script(src="/javascripts/GraphState.js")
   script(src="/javascripts/UUIDGenerator-devUseOnly.js")
