+++ views/layout.pug
@@ -7,4 +7,5 @@ html
     block content
   footer
     script(src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js")
+    script(src="/javascripts/PostOffice.js")
     block footer
